New component/file FallingCheese (bezier curve + fritasse)
